+++ tidy3d/components/simulation.py
@@ -18,9 +18,10 @@ from .geometry.mesh import TriangleMesh
 from .geometry.utils import flatten_groups, traverse_geometries
 from .geometry.utils_2d import get_bounds, set_bounds, get_thickened_geom, subdivide
 from .types import Ax, FreqBound, Axis, annotate_type, InterpMethod, Symmetry
-from .types import Literal
+from .types import Literal, TYPE_TAG_STR
 from .grid.grid import Coords1D, Grid, Coords
 from .grid.grid_spec import GridSpec, UniformGrid, AutoGrid, CustomGrid
+from .grid.grid_spec import ConformalMeshSpecType, StaircasingConformalMeshSpec
 from .medium import MediumType, AbstractMedium
 from .medium import AbstractCustomMedium, Medium2D
 from .medium import AnisotropicMedium, FullyAnisotropicMedium, AbstractPerturbationMedium
@@ -556,6 +557,13 @@ class Simulation(AbstractSimulation):
         *  `Introduction to subpixel averaging <https://www.flexcompute.com/fdtd101/Lecture-10-Introduction-to-subpixel-averaging/>`_
         *  `Dielectric constant assignment on Yee grids <https://www.flexcompute.com/fdtd101/Lecture-9-Dielectric-constant-assignment-on-Yee-grids/>`_
     """
+
+    pec_conformal_mesh_spec: ConformalMeshSpecType = pydantic.Field(
+        StaircasingConformalMeshSpec(),
+        title="Conformal mesh specifications",
+        description="Conformal mesh specifications applied to PEC strucures.",
+        discriminator=TYPE_TAG_STR,
+    )
 
     normalize_index: Union[pydantic.NonNegativeInt, None] = pydantic.Field(
         0,
@@ -2748,6 +2756,13 @@ class Simulation(AbstractSimulation):
     """ Discretization """
 
     @cached_property
+    def scaled_courant(self) -> float:
+        """When conformal mesh is applied, courant number is scaled down depending on `conformal_mesh_spec`."""
+        if self.subpixel:
+            return self.courant * self.pec_conformal_mesh_spec.courant_ratio
+        return self.courant
+
+    @cached_property
     def dt(self) -> float:
         """Simulation time step (distance).
 
@@ -2761,7 +2776,7 @@ class Simulation(AbstractSimulation):
         dl_avg = 1 / np.sqrt(dl_sum_inv_sq)
         # material factor
         n_cfl = min(min(mat.n_cfl for mat in self.scene.mediums), 1)
-        return n_cfl * self.courant * dl_avg / C_0
+        return n_cfl * self.scaled_courant * dl_avg / C_0
 
     @cached_property
     def tmesh(self) -> Coords1D: